mediacapture-streams/MediaStreamTrack-MediaElement-disabled-audio-is-silence.https.html

<!doctype html>
<title>A disabled audio track is rendered as silence</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#introduction">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#mediastreams-as-media-elements">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<p class="instructions">When prompted, accept to share your audio stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that a disabled audio track in a
MediaStream is rendered as silence. It relies on the
<a href="https://dvcs.w3.org/hg/audio/raw-file/tip/webaudio/specification.html">
Web Audio API</a>.</p>

<script>

promise_test(async t => {
  const aud = document.getElementById("aud");
  const stream = await navigator.mediaDevices.getUserMedia({audio: true});
  t.add_cleanup(() => stream.getAudioTracks()[0].stop());
  const ctx = new AudioContext();
  const streamSource = ctx.createMediaStreamSource(stream);
  const silenceDetector = ctx.createScriptProcessor(1024);
  const watcher = new EventWatcher(t, silenceDetector, ["audioprocess"]);

  stream.getAudioTracks()[0].enabled = false;
  streamSource.connect(silenceDetector);
  silenceDetector.connect(ctx.destination);

  for (let count = 10; count >= 0; count--) {
    let e = await watcher.wait_for("audioprocess");
    let buffer1 = e.inputBuffer.getChannelData(0);
    let buffer2 = e.inputBuffer.getChannelData(1);
    let out = e.outputBuffer.getChannelData(0);
    out = new Float32Array(buffer1);
    for (let i = 0; i < buffer1.length; i++) {
      assert_equals(buffer1[i], 0, "Audio buffer entry #" + i + " in channel 0 is silent");
    }
    for (let i = 0; i < buffer2.length; i++) {
      assert_equals(buffer2[i], 0, "Audio buffer entry #" + i + " in channel 1 is silent");
    }
  }
  silenceDetector.onaudioprocess = null;
}, "Tests that a disabled audio track in a MediaStream is rendered as silence");

</script>
